# apps/api/app/features/transactions/models.py
"""Pydantic v2 schemas for the transactions feature.

The schemas are the single source of truth for both the FastAPI router
(via ``response_model``) and the OpenAPI emit that backs the generated
TypeScript SDK in ``packages/client-sdk``.

Pydantic only enforces *structural* validation here — types, regex,
length / range bounds. The richer per-split-method invariants
(percent sums to 100, payers ⊆ members, owed sums match amount,
settlement shape) live in :mod:`app.features.transactions.service`
because they need to raise typed application errors with stable codes
(``OWED_SUM``, ``PERCENT_SUM``, ``PAYER_NOT_MEMBER`` …) that the
client UX depends on. Raising those from inside Pydantic validators
collapses them into a single 422 ``VALIDATION_ERROR``.
"""
from __future__ import annotations

from datetime import date, datetime
from decimal import Decimal
from typing import Annotated, Literal

from pydantic import BaseModel, ConfigDict, Field, StringConstraints

Currency = Literal["USD", "INR"]
SplitMethod = Literal["equal", "amount", "share", "percent"]
TxnType = Literal["expense", "settlement"]
SettlementStatus = Literal["settled", "friend_owes", "you_owe"]

# Crockford-base32 ULID; same regex used in the friends feature.
_ULID_PATTERN = r"^[0-9A-HJKMNP-TV-Z]{26}$"

# Domain bounds.
MIN_MEMBERS = 2
MAX_MEMBERS = 10
NAME_MAX = 120
NOTE_MAX = 500
DATE_FUTURE_TOLERANCE_DAYS = 1
DATE_PAST_HORIZON_DAYS = 365 * 10
PERCENT_TOLERANCE = Decimal("0.01")
SUM_TOLERANCE = Decimal("0.01")


# ---------------------------------------------------------------------------
# Inputs
# ---------------------------------------------------------------------------


class MemberInput(BaseModel):
    """One member entry on a create request."""

    model_config = ConfigDict(extra="forbid")

    user_id: Annotated[str, Field(pattern=_ULID_PATTERN, min_length=26, max_length=26)]
    share: (
        Annotated[
            Decimal,
            Field(gt=0, max_digits=12, decimal_places=4),
        ]
        | None
    ) = None
    percent: (
        Annotated[
            Decimal,
            Field(ge=0, le=100, max_digits=5, decimal_places=2),
        ]
        | None
    ) = None
    owed_amount: (
        Annotated[
            Decimal,
            Field(ge=0, max_digits=12, decimal_places=2),
        ]
        | None
    ) = None


class PayerInput(BaseModel):
    """One payer entry on a create request."""

    model_config = ConfigDict(extra="forbid")

    user_id: Annotated[str, Field(pattern=_ULID_PATTERN, min_length=26, max_length=26)]
    paid_amount: Annotated[
        Decimal, Field(gt=0, max_digits=12, decimal_places=2)
    ]


class CreateTransactionRequest(BaseModel):
    """``POST /v1/transactions`` request body.

    Structural validation only here — per-method invariants are in
    :func:`app.features.transactions.service.validate_create_payload`.
    """

    model_config = ConfigDict(extra="forbid")

    name: Annotated[
        str,
        StringConstraints(min_length=1, max_length=NAME_MAX, strip_whitespace=True),
    ]
    type: TxnType
    amount: Annotated[Decimal, Field(gt=0, max_digits=12, decimal_places=2)]
    currency: Currency
    txn_date: date
    note: Annotated[
        str, StringConstraints(max_length=NOTE_MAX, strip_whitespace=False)
    ] = ""
    split_method: SplitMethod
    members: Annotated[list[MemberInput], Field(min_length=1, max_length=MAX_MEMBERS)]
    payers: Annotated[list[PayerInput], Field(min_length=1, max_length=MAX_MEMBERS)]


# ---------------------------------------------------------------------------
# Outputs (response shapes)
# ---------------------------------------------------------------------------


class Member(BaseModel):
    """One member, with the server-computed ``owed_amount``."""

    user_id: str
    owed_amount: Decimal
    share: Decimal | None = None
    percent: Decimal | None = None


class Payer(BaseModel):
    user_id: str
    paid_amount: Decimal


class Transaction(BaseModel):
    """``POST /v1/transactions`` 201 + ``GET /v1/transactions/{id}`` 200."""

    txn_id: str
    creator_id: str
    name: str
    type: TxnType
    amount: Decimal
    currency: Currency
    txn_date: date
    note: str
    split_method: SplitMethod
    members: list[Member]
    payers: list[Payer]
    created_at: datetime
    updated_at: datetime
    deleted_at: datetime | None = None


class TransactionListItem(BaseModel):
    """One row in ``GET /v1/transactions`` — META + the requester's owed."""

    txn_id: str
    name: str
    type: TxnType
    amount: Decimal
    currency: Currency
    txn_date: date
    split_method: SplitMethod
    creator_id: str
    my_owed_amount: Decimal
    created_at: datetime


MAX_LIST_LIMIT = 100
DEFAULT_LIST_LIMIT = 20


class ListTransactionsQuery(BaseModel):
    model_config = ConfigDict(extra="forbid")

    limit: Annotated[int, Field(ge=1, le=MAX_LIST_LIMIT)] = DEFAULT_LIST_LIMIT
    cursor: str | None = None
    friend_id: (
        Annotated[
            str, Field(pattern=_ULID_PATTERN, min_length=26, max_length=26)
        ]
        | None
    ) = None


class ListTransactionsResponse(BaseModel):
    items: list[TransactionListItem]
    next_cursor: str | None = None


__all__ = [
    "DATE_FUTURE_TOLERANCE_DAYS",
    "DATE_PAST_HORIZON_DAYS",
    "DEFAULT_LIST_LIMIT",
    "MAX_LIST_LIMIT",
    "MAX_MEMBERS",
    "MIN_MEMBERS",
    "PERCENT_TOLERANCE",
    "SUM_TOLERANCE",
    "CreateTransactionRequest",
    "Currency",
    "ListTransactionsQuery",
    "ListTransactionsResponse",
    "Member",
    "MemberInput",
    "Payer",
    "PayerInput",
    "SettlementStatus",
    "SplitMethod",
    "Transaction",
    "TransactionListItem",
    "TxnType",
]